+++ cppython/utility/exception.py
@@ -35,3 +35,21 @@ class PluginError(Exception):
             str -- The underlying error
         """
         return self._error
+
+
+class NotSupportedError(Exception):
+    """Raised when something is not supported"""
+
+    def __init__(self, error: str) -> None:
+        self._error = error
+
+        super().__init__(error)
+
+    @property
+    def error(self) -> str:
+        """Returns the underlying error
+
+        Returns:
+            str -- The underlying error
+        """
+        return self._error
