+++ tests/unit/core/test_resolution.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 from pydantic import Field
 
-from cppython.core.exceptions import ConfigException
+from cppython.core.exception import ConfigException
 from cppython.core.plugin_schema.generator import Generator
 from cppython.core.plugin_schema.provider import Provider
 from cppython.core.plugin_schema.scm import SCM
